Big live update, see details
Import the code for network live reading.
Split the code for mmap and network live in separate files.
Disable by default the support for mmap live reading, needs to be
defined manually with -DLTTNGTOP_MMAP_LIVE for now, since the libraries it
depends on are not maintained.
Sync the entire babeltrace include folder.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>