lttngtop without arguments starts a local live trace
Get rid of the code trying to attach to a live mmap session, since the
support for this mode of operation is not going to enter anytime soon in
lttng-tools and babeltrace.

Instead, we try to create a live session on localhost with all the
required parameters and try to attach to it.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>